+++ src/trace_processor/perfetto_sql/stdlib/android/frames/per_frame_metrics.sql
@@ -26,10 +26,10 @@ INCLUDE PERFETTO MODULE android.frames.timeline;
 -- For Googlers: more details in go/android-performance-metrics-glossary.
 CREATE PERFETTO TABLE android_frames_overrun(
     -- Frame id.
-    frame_id INT,
+    frame_id LONG,
     -- Difference between `expected` and `actual` frame ends. Negative if frame
     -- didn't miss deadline.
-    overrun INT
+    overrun LONG
 ) AS
 SELECT
     frame_id,
@@ -42,9 +42,9 @@ JOIN slice exp_slice ON (exp.id = exp_slice.id);
 -- How much time did the frame's Choreographer callbacks take.
 CREATE PERFETTO TABLE android_frames_ui_time(
     -- Frame id
-    frame_id INT,
+    frame_id LONG,
     -- UI time duration
-    ui_time INT
+    ui_time LONG
 ) AS
 SELECT
     frame_id,
@@ -60,9 +60,9 @@ JOIN slice USING (id);
 -- For Googlers: more details in go/android-performance-metrics-glossary.
 CREATE PERFETTO TABLE android_app_vsync_delay_per_frame(
     -- Frame id
-    frame_id INT,
+    frame_id LONG,
     -- App VSYNC delay.
-    app_vsync_delay INT
+    app_vsync_delay LONG
 ) AS
 -- As there can be multiple `DrawFrame` slices, the `frames_surface_slices`
 -- table contains multiple rows for the same `frame_id` which only differ on
@@ -92,17 +92,17 @@ JOIN slice act ON (f.actual_frame_timeline_id = act.id);
 -- For Googlers: more details in go/android-performance-metrics-glossary.
 CREATE PERFETTO TABLE android_cpu_time_per_frame(
     -- Frame id
-    frame_id INT,
+    frame_id LONG,
     -- Difference between actual timeline of the frame and
     -- `Choreographer#doFrame`. See `android_app_vsync_delay_per_frame` table for more details.
-    app_vsync_delay INT,
+    app_vsync_delay LONG,
     -- Duration of `Choreographer#doFrame` slice.
-    do_frame_dur INT,
+    do_frame_dur LONG,
     -- Duration of `DrawFrame` slice. Summed duration of all `DrawFrame`
     -- slices, if more than one. See `android_frames_draw_frame` for more details.
-    draw_frame_dur INT,
+    draw_frame_dur LONG,
     -- CPU time across the UI Thread + RenderThread.
-    cpu_time INT
+    cpu_time LONG
 ) AS
 WITH all_draw_frames AS (
 SELECT
@@ -136,9 +136,9 @@ JOIN slice do_frame ON (f.do_frame_id = do_frame.id);
 -- For Googlers: more details in go/android-performance-metrics-glossary.
 CREATE PERFETTO TABLE _cpu_time_per_frame_fallback(
     -- Frame id.
-    frame_id INT,
+    frame_id LONG,
     -- Estimated cpu time.
-    estimated_cpu_time INT
+    estimated_cpu_time LONG
 ) AS
 SELECT
     frame_id,
@@ -146,8 +146,8 @@ SELECT
 FROM android_frames_ui_time;
 
 CREATE PERFETTO TABLE _estimated_cpu_time_per_frame(
-    frame_id INT,
-    cpu_time INT
+    frame_id LONG,
+    cpu_time LONG
 ) AS
 SELECT
     frame_id,
@@ -160,14 +160,14 @@ LEFT JOIN android_cpu_time_per_frame r USING (frame_id);
 -- For Googlers: more details in go/android-performance-metrics-glossary.
 CREATE PERFETTO TABLE android_frame_stats(
     -- Frame id.
-    frame_id INT,
+    frame_id LONG,
     -- The amount by which each frame missed of hit its deadline. See
     -- `android_frames_overrun` for details.
-    overrun INT,
+    overrun LONG,
     -- How much time did the frame take across the UI Thread + RenderThread.
-    cpu_time INT,
+    cpu_time LONG,
     -- How much time did the frame's Choreographer callbacks take.
-    ui_time INT,
+    ui_time LONG,
     -- Was frame janky.
     was_jank BOOL,
     -- CPU time of the frame took over 20ms.
